Models CRUD

+++ back/src/main/scala/phystech/data/Model.scala
@@ -1,8 +1,10 @@
 package phystech.data
 
-final case class Model(modelId: String,
-                       parentModelId: String,
-                       modelName: String,
-                       version: Int,
-                       testingStage: Int,
-                       variableList: List[ModelVariable])
+final case class Model(
+  modelId: String,
+  parentModelId: String,
+  modelName: String,
+  version: Long,
+  testingStage: Int,
+  variableList: List[ModelVariable]
+)
